refactor: use score-helm in cmd directory
